# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/accounts/balance": {
      "get": {
        "description": "Récupère le solde courant pour toutes les devises ou pour une devise précise",
        "operationId": "AccountsController_getBalance",
        "parameters": [
          {
            "name": "currency",
            "required": false,
            "in": "query",
            "description": "Filtrer par code devise (XOF, USD, EUR)",
            "schema": {
              "enum": [
                "XOF",
                "USD",
                "EUR"
              ],
              "type": "string"
            }
          },
          {
            "name": "Lomi-Version",
            "in": "header",
            "required": false,
            "description": "Optional schema version pin. Echoes OpenAPI info.version (currently 1.2.0). Routes stay unversioned.",
            "schema": {
              "type": "string",
              "example": "1.2.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Informations de solde",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountBalanceResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Solde du compte",
        "tags": [
          "Balances"
        ]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "Payment processing API for francophone West African businesses. Routes stay unversioned; OpenAPI info.version is the schema release. Send optional Lomi-Version to pin a known schema. Compatibility is additive: existing fields are not removed without a major info.version bump.",
    "version": "1.2.0",
    "contact": {
      "name": "lomi.",
      "url": "https://docs.lomi.africa",
      "email": "hello@lomi.africa"
    }
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      },
      "provisioning-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Lomi-Provisioning-Key"
      },
      "partner-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Lomi-Partner-Key"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.lomi.africa/oauth/authorize",
            "tokenUrl": "https://api.lomi.africa/oauth/token",
            "refreshUrl": "https://api.lomi.africa/oauth/token",
            "scopes": {
              "provisioning.onboard": "Create and complete merchant onboarding",
              "merchant.read": "Read merchant resources",
              "merchant.write": "Write merchant resources (no payouts, refunds, or instant settlement)",
              "merchant.money": "Move money: payouts, refunds, and instant settlement"
            }
          }
        },
        "description": "Agent OAuth for MCP and Connect. Merchant REST still uses X-API-KEY by default."
      }
    },
    "schemas": {
      "AccountBalanceResponseDto": {
        "type": "object",
        "properties": {
          "currency_code": {
            "type": "string",
            "example": "XOF",
            "description": "Currency code"
          },
          "balance": {
            "type": "number",
            "example": 100000,
            "description": "Account balance"
          },
          "last_updated": {
            "type": "string",
            "example": "2024-01-01T00:00:00Z",
            "description": "Last updated timestamp"
          }
        },
        "required": [
          "currency_code",
          "balance",
          "last_updated"
        ]
      },
...
    },
...
  }
}